Commit migrated from https://github.com/dotnet/coreclr/commit/f2efef85bd811169ccc2e1896fb9d689b1fb459b

+++ src/coreclr/tests/scripts/run-corefx-tests.py
@@ -17,6 +17,7 @@
 ##########################################################################
 
 import argparse
+import distutils.dir_util
 import os
 import re
 import shutil
@@ -63,6 +64,7 @@ parser.add_argument('-build_type', dest='build_type', default='Debug')
 parser.add_argument('-clr_root', dest='clr_root', default=None)
 parser.add_argument('-fx_root', dest='fx_root', default=None)
 parser.add_argument('-fx_branch', dest='fx_branch', default='master')
+parser.add_argument('-fx_commit', dest='fx_commit', default=None)
 parser.add_argument('-env_script', dest='env_script', default=None)
 
 
@@ -75,8 +77,8 @@ def validate_args(args):
     Args:
         args (argparser.ArgumentParser): Args parsed by the argument parser.
     Returns:
-        (arch, build_type, clr_root, fx_root, fx_branch, env_script)
-            (str, str, str, str, str, str)
+        (arch, build_type, clr_root, fx_root, fx_branch, fx_commit, env_script)
+            (str, str, str, str, str, str, str)
     Notes:
     If the arguments are valid then return them all in a tuple. If not, raise
     an exception stating x argument is incorrect.
@@ -87,6 +89,7 @@ def validate_args(args):
     clr_root = args.clr_root
     fx_root = args.fx_root
     fx_branch = args.fx_branch
+    fx_commit = args.fx_commit
     env_script = args.env_script
 
     def validate_arg(arg, check):
@@ -115,6 +118,9 @@ def validate_args(args):
     validate_arg(build_type, lambda item: item in valid_build_types)
     validate_arg(fx_branch, lambda item: True)
 
+    if fx_commit is None:
+        fx_commit = 'HEAD'
+
     if clr_root is None:
         clr_root = nth_dirname(os.path.abspath(sys.argv[0]), 3)
     else:
@@ -130,7 +136,7 @@ def validate_args(args):
         validate_arg(env_script, lambda item: os.path.isfile(env_script))
         env_script = os.path.abspath(env_script)
 
-    args = (arch, build_type, clr_root, fx_root, fx_branch, env_script)
+    args = (arch, build_type, clr_root, fx_root, fx_branch, fx_commit, env_script)
 
     log('Configuration:')
     log(' arch: %s' % arch)
@@ -138,10 +144,10 @@ def validate_args(args):
     log(' clr_root: %s' % clr_root)
     log(' fx_root: %s' % fx_root)
     log(' fx_branch: %s' % fx_branch)
+    log(' fx_commit: %s' % fx_commit)
     log(' env_script: %s' % env_script)
 
     return args
-
 
 def nth_dirname(path, n):
     """ Find the Nth parent directory of the given path
@@ -182,7 +188,6 @@ def log(message):
 
     print '[%s]: %s' % (sys.argv[0], message)
 
-
 ##########################################################################
 # Main
 ##########################################################################
@@ -191,7 +196,9 @@ def main(args):
     global Corefx_url
     global Unix_name_map
 
-    arch, build_type, clr_root, fx_root, fx_branch, env_script = validate_args(
+    testing = False
+
+    arch, build_type, clr_root, fx_root, fx_branch, fx_commit, env_script = validate_args(
         args)
 
     clr_os = 'Windows_NT' if Is_windows else Unix_name_map[os.uname()[0]]
@@ -206,81 +213,112 @@ def main(args):
     # To delete the files with non-ascii characters, when rmtree fails due to those
     # files, we then will call rd on Windows.
 
-    if os.path.exists(fx_root):
+    if not testing and os.path.exists(fx_root):
         if Is_windows:
-            vbcscompiler_running = True
-            while vbcscompiler_running:
+            while True:
                 res = subprocess.check_output(['tasklist'])
                 if not 'VBCSCompiler.exe' in res:
-                    vbcscompiler_running = False
+                   break                
         os.chdir(fx_root)
         os.system('git clean -fxd')
         os.chdir(clr_root)
         shutil.rmtree(fx_root, onerror=del_rw)
 
+    # Clone the corefx branch
+
     command = 'git clone -b %s --single-branch %s %s' % (
         fx_branch, Corefx_url, fx_root)
-
     log(command)
-
-    testing = False
-
     if testing:
-        os.makedirs(fx_root)
+        if not os.path.exists(fx_root):
+            os.makedirs(fx_root)
         returncode = 0
     else:
         returncode = os.system(command)
 
-    if returncode != 0:
-        sys.exit(returncode)
+    # Change directory to the corefx root
 
     cwd = os.getcwd()
-    log('cd ' + fx_root)
+    log('[cd] ' + fx_root)
     os.chdir(fx_root)
 
-    if Is_windows:
-        command = '.\\build.cmd'
-        if env_script is not None:
-            command = ('cmd /c %s&&' % env_script) + command
-    else:
-        # CoreFx build.sh requires HOME to be set, and it isn't by default
-        # under our CI.
+    # Checkout the appropriate corefx commit
+
+    command = "git checkout %s" % fx_commit
+    log(command)
+    returncode = 0 if testing else os.system(command)
+    if not returncode == 0:
+        sys.exit(returncode)
+
+    # On Unix, coreFx build.sh requires HOME to be set, and it isn't by default
+    # under our CI system, so set it now.
+
+    if not Is_windows:
         fx_home = os.path.join(fx_root, 'tempHome')
         if not os.path.exists(fx_home):
             os.makedirs(fx_home)
         os.putenv('HOME', fx_home)
         log('HOME=' + fx_home)
 
-        command = './build.sh'
-        if env_script is not None:
-            command = ('. %s;' % env_script) + command
-
+    # Determine the RID to specify the to corefix build scripts.  This seems to
+    # be way harder than it ought to be.
+ 
     if testing:
         rid_os = dotnet_rid_os('')
     else:
-        if clr_os == "Windows_NT":
+        if Is_windows:
             rid_os = "win7"
         else:
             rid_os = dotnet_rid_os(os.path.join(clr_root, 'Tools', 'dotnetcli'))
 
+    # Gather up some arguments to pass to both build and build-tests.
+
+    config_args = '-Release -RuntimeOS=%s -ArchGroup=%s' % (rid_os, arch)
+
+    # Run the primary (non-test) corefx build
+
+    command = ' '.join(('build.cmd' if Is_windows else './build.sh', config_args))
+    log(command)
+    returncode = 0 if testing else os.system(command)
+    if returncode != 0:
+        sys.exit(returncode)
+
+    # Copy the coreclr runtime we wish to run tests against.  This is the recommended
+    # hack until a full-stack test solution is ready.  This assumes there is a single
+    # directory under <fx_root>/bin/runtime into which we copy coreclr binaries.  We
+    # assume the appropriate coreclr has already been built.
+
+    fx_runtime_dir = os.path.join(fx_root, 'bin', 'runtime')
+    overlay_dest = os.path.join(fx_runtime_dir, os.listdir(fx_runtime_dir)[0])
+    log('[overlay] %s -> %s' % (core_root, overlay_dest))
+    if not testing:
+        distutils.dir_util.copy_tree(core_root, overlay_dest)
+
+    # Build the build-tests command line.
+
+    if Is_windows:
+        command = 'build-tests.cmd'
+        if env_script is not None:
+            command = ('cmd /c %s&&' % env_script) + command
+    else:
+        command = './build-tests.sh'
+        if env_script is not None:
+            command = ('. %s;' % env_script) + command
+
     command = ' '.join((
         command,
-        '-Release',
-        '-TestNugetRuntimeId=%s-%s' % (rid_os, arch),
+        config_args,
         '--',
-        '/p:BUILDTOOLS_OVERRIDE_RUNTIME="%s"' % core_root,
         '/p:WithoutCategories=IgnoreForCI'
     ))
 
     if not Is_windows:
         command += ' /p:TestWithLocalNativeLibraries=true'
 
-    log(command)
+    # Run the corefx test build and run the tests themselves.
 
-    if testing:
-        returncode = 0
-    else:
-        returncode = os.system(command)
+    log(command)
+    returncode = 0 if testing else os.system(command)
 
     sys.exit(returncode)
